fix readJson() file existence check + improve logging
